Equitably Allocate Vents Among Hospitals.
The general idea of the code is that if you have patient zip code of residence, you can use the racial and socioeconomic makeup of that zip code to impute the demographic makeup of the hospital’s patients. We later compared this to self reported racial data, and found that it correlated favorably.